Extract common functions to independent package
